Moved top nav and drawer
